pyproject.toml
[project]
name = "AREMA-ontology"
version = "0.1.0"
description = "Ontology tooling and docs generation for AREMA"
authors = [{ name = "Robin Franken", email = "robin.franken@epfl.ch" }]
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "rdflib==7.1.4",
    "pyshacl==0.30.1",
    "pandas>=2.1.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "apscheduler>=3.10.0",
    "google-api-python-client>=2.187.0",
    "google-auth>=2.45.0",
    "gitpython>=3.1.45",
    "inflection>=0.5.1",
]

[tool.uv]
